Move devices import to top level
